requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Authentication
Flask-Login==0.6.3
bcrypt==4.1.2

# Forms & CSRF
Flask-WTF==1.2.1
WTForms==3.1.1

# Markdown Processing
markdown==3.5.1
bleach==6.1.0

# Image Processing
Pillow==10.1.0

# HTTP Requests (for GitHub API)
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0

# PostgreSQL Support (optional, for production)
psycopg2-binary==2.9.9